refactor: extract scar weakness section
